migrate sql: from Integer to BigInteger

+++ zeldris/modules/sql/connection_sql.py
@@ -20,7 +20,7 @@ import time
 from typing import Union
 
 from sqlalchemy import Column, String, Boolean, UnicodeText, Integer
-
+from sqlalchemy.sql.sqltypes import BigInteger
 from zeldris.modules.sql import SESSION, BASE
 
 
@@ -41,7 +41,7 @@ class ChatAccessConnectionSettings(BASE):
 
 class Connection(BASE):
     __tablename__ = "connection"
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     chat_id = Column(String(14))
 
     def __init__(self, user_id, chat_id):
@@ -51,7 +51,7 @@ class Connection(BASE):
 
 class ConnectionHistory(BASE):
     __tablename__ = "connection_history"
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     chat_id = Column(String(14), primary_key=True)
     chat_name = Column(UnicodeText)
     conn_time = Column(Integer)
